Deduplicate payjoin URI creation logic
Create a helper function `create_pj_uri` in the v2 module that centralizes
the logic for creating Payjoin URIs from a SessionContext. This eliminates
code duplication between `SessionHistory::pj_uri` and
`Receiver<WithContext>::pj_uri`.

The helper function takes a SessionContext reference and an OutputSubstitution
parameter, making it flexible for both use cases while maintaining the same
behavior as the original implementations.

This change:
- Improves maintainability by centralizing URI creation logic
- Reduces code duplication
- Preserves the existing API and behavior
